+++ stockfish-core/CHANGELOG.md
@@ -24,6 +24,7 @@ adheres to [Semantic Versioning][semver].
 - `impl IntoIterator<Output = Square> for Bitboard`
 - `impl BitOr<Color, Output = Piece> for Token`
 - `impl BitOr<Token, Output = Piece> for Color`
+- `Token::attacks()` and `Token::moves()`
 
 ## [0.2.0](https://github.com/stouset/stockfish-rs/tree/stockfish-core-v0.2.0/stockfish-core) - 2023-01-16